+++ Config.cmake/test_fmi1.cmake
@@ -44,6 +44,7 @@ set(XML_ME_PATH ${FMU_DUMMY_FOLDER}/modelDescription_me.xml)
 set(XML_CS_PATH ${FMU_DUMMY_FOLDER}/modelDescription_cs.xml)
 set(XML_CS_TC_PATH ${FMU_DUMMY_FOLDER}/modelDescription_cs_tc.xml)
 set(XML_MF_PATH ${FMU_DUMMY_FOLDER}/modelDescription_malformed.xml)
+set(TYPE_DEFINITIONS_MODEL_DESC_DIR ${RTTESTDIR}/FMI1/parser_test_xmls/type_definitions)
 
 set(SHARED_LIBRARY_ME_PATH ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_CFG_INTDIR}/${CMAKE_SHARED_LIBRARY_PREFIX}fmu1_dll_me${CMAKE_SHARED_LIBRARY_SUFFIX})
 set(SHARED_LIBRARY_CS_PATH ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_CFG_INTDIR}/${CMAKE_SHARED_LIBRARY_PREFIX}fmu1_dll_cs${CMAKE_SHARED_LIBRARY_SUFFIX})
@@ -66,6 +67,8 @@ compress_fmu("${TEST_OUTPUT_FOLDER}" "${FMU_DUMMY_CS_MODEL_IDENTIFIER}" "cs_tc" 
 
 add_executable (fmi1_import_default_experiment_test ${RTTESTDIR}/FMI1/fmi1_import_default_experiment_test.c)
 target_link_libraries (fmi1_import_default_experiment_test  ${FMILIBFORTEST}  )
+add_executable(fmi1_type_definitions_test ${RTTESTDIR}/FMI1/fmi1_import_type_definitions_test.c)
+target_link_libraries(fmi1_type_definitions_test ${FMILIBFORTEST})
 
 add_executable (fmi1_xml_parsing_test ${RTTESTDIR}/FMI1/fmi1_xml_parsing_test.c)
 target_link_libraries (fmi1_xml_parsing_test  ${FMILIBFORTEST}  )
@@ -88,6 +91,7 @@ to_native_c_path(${TEST_OUTPUT_FOLDER}/tempfolder FMU_TEMPFOLDER)
 
 add_test(ctest_fmi1_xml_parsing_test fmi1_import_default_experiment_test ${RTTESTDIR}/FMI1/parser_test_xmls/default_experiment/)
 add_test(ctest_fmi1_xml_parsing_test fmi1_xml_parsing_test ${RTTESTDIR}/FMI1/parser_test_xmls/)
+add_test(ctest_fmi1_type_definitions_test fmi1_type_definitions_test ${TYPE_DEFINITIONS_MODEL_DESC_DIR})
 ADD_TEST(ctest_fmi_import_me_test fmi_import_me_test ${FMU_ME_PATH} ${FMU_TEMPFOLDER})
 ADD_TEST(ctest_fmi_import_cs_test fmi_import_cs_test ${FMU_CS_PATH} ${FMU_TEMPFOLDER} "modelDescription_cs.xml")
 ADD_TEST(ctest_fmi_import_cs_tc_test fmi_import_cs_test ${FMU_CS_TC_PATH} ${FMU_TEMPFOLDER} "modelDescription_cs_tc.xml")
@@ -137,6 +141,7 @@ if(FMILIB_BUILD_BEFORE_TESTS)
 		ctest_fmi1_capi_me_test
 		ctest_fmi1_logger_test_run
 		ctest_fmi1_xml_parsing_test
+        ctest_fmi1_type_definitions_test
 		PROPERTIES DEPENDS ctest_build_all)
 endif()